moved implementation of analysis methods on Ss struct to analysis module
